feat: server-render article pages with ISR (fixes #207)
- articles/[id] is now an async server component: article HTML is in the
  initial response (no client fetch waterfall), with ISR (revalidate 300s)
  and real metadata/OpenGraph for SEO. No generateStaticParams: the backend
  isn't reachable during docker build, so pages render on first request.
- articles list is a server component (force-dynamic to avoid build-time
  prerender; article list still cached via the Next data cache) passing
  data to the existing interactive client UI (ArticlesClient).
- Client islands: AIAssistant (panel+button+feature flags), RelatedNotes,
  ArticleTags, ArticleTableOfContents.
- sanitize.ts now uses isomorphic-dompurify so backend HTML is sanitized
  on the server too (previous code skipped sanitization when window was
  undefined); kept external via serverExternalPackages (jsdom reads its
  default stylesheet from disk and breaks when bundled).
- API_URL_INTERNAL (http://backend:8000) added to dev/prod compose for
  server-side fetches; browser keeps NEXT_PUBLIC_API_URL.
- Removed the /api/articles/{id} hover prefetch: the detail page no longer
  client-fetches, and Next prefetches the RSC payload itself.
- Fixed detail-page tag links pointing at ?tag= while the list reads ?tags=.
