+++ frontend/package.json
@@ -30,6 +30,7 @@
     "@uiw/react-codemirror": "^4.25.2",
     "d3": "^7.9.0",
     "dompurify": "^3.3.1",
+    "isomorphic-dompurify": "^3.18.0",
     "next": "^15.5.7",
     "react": "^19.2.1",
     "react-dom": "^19.2.1",
